Remake lock_pairs() func
Added logic in the lock_pairs function to check whether locking
a given pair would introduce a cycle in the directed graph. This
ensures the graph remains acyclic, which is necessary for the
Tideman method to correctly determine the election winner.

detect_cycle() will be added in the next commit